Function : 전사 행동 구현 (4)
- 스탯을 모든 행동에 연결해둠
- Player의 데이터 테이블과 관련된 Key값을 enum에서 int로 변경

+++ Assets/3_Scripts/Managers/TableMgr.cs
@@ -44,8 +44,8 @@ public class TableMgr
 
     public void LinkWarriorTable()
     {
-        character.InitPlayerStatTableCsv("WarriorStatTable", 1, 0, PlayerEnums.TYPEIDS.WARRIOR);
-        character.InitPlayerNormalAttackTableCsv("WarriorNormalAttackTable", 1, 0, PlayerEnums.TYPEIDS.WARRIOR, 3);
+        character.InitPlayerStatTableCsv("WarriorStatTable", 1, 0, (int)PlayerEnums.TYPEIDS.WARRIOR);
+        character.InitPlayerNormalAttackTableCsv("WarriorNormalAttackTable", 1, 0, (int)PlayerEnums.TYPEIDS.WARRIOR, 3);
         character.InitPlayerBuffSkillTableCsv("WarriorBuffSkillTable", 1, 0, PlayerEnums.BUFF_SKILLS.WARRIOR_ROAR, 2);
         character.InitPlayerAttackSkillTableCsv("WarriorAttackUltimateSkillTable", 1, 0, PlayerEnums.ATTACK_SKILLS.WARRIOR_ULTIMATE);
     }
